This is a big change, and I'm sure there will still be bugs to be worked out. I
changed to [3]float64 because of two reasons:

1) Using arrays simplifies a lot of code without losing the by-value passing.
2) I want to use the highest precision for geometry.

The code compiles and works as-is, but as with everything, I'm sure there are
still places where I can be more efficient.

Another change is the addition of the Axis type. This should a nice management
relief, as there are many "axis" parameters that were just being stored as ints.

+++ src/goray/core/camera/camera.go
@@ -17,13 +17,13 @@ type Camera interface {
 	// ShootRay calculates the initial ray used for computing a fragment of the
 	// output.  U and V are sample coordinates that are only calculated if
 	// SampleLens returns true.
-	ShootRay(x, y, u, v float) (ray.Ray, float)
+	ShootRay(x, y, u, v float64) (ray.Ray, float64)
 	// ResolutionX returns the number of fragments wide that the camera is.
 	ResolutionX() int
 	// ResolutionY returns the number of fragments high that the camera is.
 	ResolutionY() int
 	// Project calculates the projection of a ray onto the fragment plane.
-	Project(wo ray.Ray, lu, lv *float) (pdf float, changed bool)
+	Project(wo ray.Ray, lu, lv *float64) (pdf float64, changed bool)
 	// SampleLens returns whether the lens needs to be sampled using the u and v
 	// parameters of ShootRay.  This is useful for DOF-like effects.  When this
 	// returns false, no lens samples need to be computed.
