+++ src/goray/std/cameras/ortho.go
@@ -24,7 +24,7 @@ type orthoCam struct {
 }
 
 // NewOrtho creates a new orthographic camera.
-func New(pos, look, up vector.Vector3D, resx, resy int, aspect, scale float) camera.Camera {
+func New(pos, look, up vector.Vector3D, resx, resy int, aspect, scale float64) camera.Camera {
 	c := new(orthoCam)
 	c.resx, c.resy = resx, resy
 	c.vup = vector.Sub(up, pos)
@@ -37,12 +37,12 @@ func New(pos, look, up vector.Vector3D, resx, resy int, aspect, scale float) cam
 	c.vright = c.vright.Normalize()
 
 	c.vright = vector.ScalarMul(c.vright, -1.0)
-	c.vup = vector.ScalarMul(c.vup, aspect*float(resy)/float(resx))
+	c.vup = vector.ScalarMul(c.vup, aspect*float64(resy)/float64(resx))
 
 	c.position = vector.Sub(pos, vector.ScalarMul(vector.Add(c.vup, c.vright), 0.5*scale))
 
-	c.vup = vector.ScalarMul(c.vup, scale/float(resy))
-	c.vright = vector.ScalarMul(c.vright, scale/float(resx))
+	c.vup = vector.ScalarMul(c.vup, scale/float64(resy))
+	c.vright = vector.ScalarMul(c.vright, scale/float64(resx))
 	return c
 }
 
@@ -50,7 +50,7 @@ func (c *orthoCam) SampleLens() bool { return false }
 func (c *orthoCam) ResolutionX() int { return c.resx }
 func (c *orthoCam) ResolutionY() int { return c.resy }
 
-func (c *orthoCam) ShootRay(x, y, u, v float) (r ray.Ray, wt float) {
+func (c *orthoCam) ShootRay(x, y, u, v float64) (r ray.Ray, wt float64) {
 	wt = 1
 	r = ray.Ray{
 		From: vector.Add(c.position, vector.ScalarMul(c.vright, x), vector.ScalarMul(c.vup, y)),
@@ -60,7 +60,7 @@ func (c *orthoCam) ShootRay(x, y, u, v float) (r ray.Ray, wt float) {
 	return
 }
 
-func (c *orthoCam) Project(wo ray.Ray, lu, lv *float) (pdf float, changed bool) {
+func (c *orthoCam) Project(wo ray.Ray, lu, lv *float64) (pdf float64, changed bool) {
 	return 0.0, false
 }
 
@@ -90,6 +90,6 @@ func Construct(m yamldata.Map) (data interface{}, err os.Error) {
 		scale = 1.0
 	}
 	// Create camera (finally!)
-	data = New(pos, look, up, int(width), int(height), float(aspect), float(scale))
+	data = New(pos, look, up, int(width), int(height), aspect, scale)
 	return
 }
